Separate SWT and non SWT in PlacementUtil.
